+++ website/app/bootie.ts
@@ -1,0 +1,5 @@
+
+export interface Bootie {
+  id: number,
+  name: string
+}
